Protego v0.1.0 Protego.Router

Handles routing depending on enabled modules.

Handles initial configuration of the enabled modules through the Protego.Router.protego_for/2 macro

Available modules

authenticatable

Functions

module_name(module)

returns the module name extracted from full module name in underscore format

Examples

iex> Protego.Router.module_name(Protego.Router)
"router"
iex> Protego.Router.module_name(Protego.RouterTest)
"router_test"
iex> Protego.Router.module_name(Protego)
"protego"

Macros

protego_for(resource, modules \\ [])

Enables protego for a specific resource with selected modules.

Examples

protego_for App.User, :authenticatable # Enables protego for App.User with the authenticatable module